+++ spec/compileOnly_spec.rb
@@ -34,6 +34,21 @@ describe "compileOnly" do
     expect(ExitHelper.exit_code).to be == 0
   end
 
+  it 'linkDirectly' do
+    Bake.startBake("compileOnly/main", ["test_ld", "-v2"])
+
+    expect($mystring.include?("g++ -c -MD -MF build/test2_ld_main_test_ld/src/lib2.d -o build/test2_ld_main_test_ld/src/lib2.o src/lib2.cpp")).to be == true
+    expect($mystring.include?("No source files, library won't be created")).to be == true
+    expect($mystring.include?("g++ -c -MD -MF build/test1_ld_main_test_ld/src/lib1a.d -o build/test1_ld_main_test_ld/src/lib1a.o src/lib1a.cpp")).to be == true
+    expect($mystring.include?("g++ -c -MD -MF build/test1_ld_main_test_ld/src/lib1b.d -o build/test1_ld_main_test_ld/src/lib1b.o src/lib1b.cpp")).to be == true
+    expect($mystring.include?("ar -rc build/test1_ld_main_test_ld/libmain.a build/test1_ld_main_test_ld/src/lib1b.o")).to be == true
+    expect($mystring.include?("g++ -c -MD -MF build/test_ld/src/lib.d -o build/test_ld/src/lib.o src/lib.cpp")).to be == true
+    expect($mystring.include?("g++ -c -MD -MF build/test_ld/src/main.d -o build/test_ld/src/main.o src/main.cpp")).to be == true
+    expect($mystring.include?("g++ -o build/test_ld/main.exe build/test_ld/src/main.o build/test_ld/src/lib.o build/test1_ld_main_test_ld/src/lib1a.o build/test1_ld_main_test_ld/libmain.a build/test2_ld_main_test_ld/src/lib2.o")).to be == true
+
+    expect(ExitHelper.exit_code).to be == 0
+  end
+
 end
 
 end